+++ daemon/cluster/cluster.go
@@ -52,8 +52,8 @@ import (
 	"github.com/containerd/log"
 	"github.com/docker/docker/daemon/cluster/controllers/plugin"
 	executorpkg "github.com/docker/docker/daemon/cluster/executor"
+	"github.com/docker/docker/daemon/internal/stack"
 	lncluster "github.com/docker/docker/daemon/libnetwork/cluster"
-	"github.com/docker/docker/pkg/stack"
 	"github.com/moby/moby/api/types/network"
 	types "github.com/moby/moby/api/types/swarm"
 	swarmapi "github.com/moby/swarmkit/v2/api"
