+++ daemon/cluster/swarm.go
@@ -9,10 +9,10 @@ import (
 
 	"github.com/containerd/log"
 	"github.com/docker/docker/daemon/cluster/convert"
+	"github.com/docker/docker/daemon/internal/stack"
 	"github.com/docker/docker/daemon/pkg/opts"
 	"github.com/docker/docker/daemon/server/backend"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/pkg/stack"
 	"github.com/moby/moby/api/types/container"
 	"github.com/moby/moby/api/types/filters"
 	types "github.com/moby/moby/api/types/swarm"
