+++ daemon/debugtrap_windows.go
@@ -7,7 +7,7 @@ import (
 	"unsafe"
 
 	"github.com/containerd/log"
-	"github.com/docker/docker/pkg/stack"
+	"github.com/docker/docker/daemon/internal/stack"
 	"golang.org/x/sys/windows"
 )
