+++ daemon/internal/stack/stackdump.go
